requirements.txt
bittensor==6.9.3
loguru
numpy
pandas
pandas_market_calendars
pydantic
pytz
rich
scikit_learn
setuptools
starlette
ta
template
tensorflow==2.15.0
torch
transformers
yfinance==0.2.17
statsmodels
huggingface_hub==0.22.2
wandb==0.16.6
python_dotenv
psycopg
shap